pyproject.toml
[project]
name = "openai-metrics"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = ["requests", "pyyaml", "openai>=1.0.0"]
